Move components.
